+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/MultiplotCollapsed.Designer.cs
@@ -1,14 +1,14 @@
-﻿namespace Sandbox.WinForms;
+﻿namespace WinForms_Demo.Demos;
 
-partial class Form1
+partial class MultiplotCollapsed
 {
     /// <summary>
-    ///  Required designer variable.
+    /// Required designer variable.
     /// </summary>
     private System.ComponentModel.IContainer components = null;
 
     /// <summary>
-    ///  Clean up any resources being used.
+    /// Clean up any resources being used.
     /// </summary>
     /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
     protected override void Dispose(bool disposing)
@@ -23,8 +23,8 @@ partial class Form1
     #region Windows Form Designer generated code
 
     /// <summary>
-    ///  Required method for Designer support - do not modify
-    ///  the contents of this method with the code editor.
+    /// Required method for Designer support - do not modify
+    /// the contents of this method with the code editor.
     /// </summary>
     private void InitializeComponent()
     {
@@ -37,19 +37,17 @@ partial class Form1
         formsPlot1.Dock = DockStyle.Fill;
         formsPlot1.Location = new Point(0, 0);
         formsPlot1.Name = "formsPlot1";
-        formsPlot1.Size = new Size(669, 415);
-        formsPlot1.TabIndex = 0;
+        formsPlot1.Size = new Size(800, 450);
+        formsPlot1.TabIndex = 7;
         // 
-        // Form1
+        // MultiplotCollapsed
         // 
         AutoScaleDimensions = new SizeF(7F, 15F);
         AutoScaleMode = AutoScaleMode.Font;
-        ClientSize = new Size(669, 415);
+        ClientSize = new Size(800, 450);
         Controls.Add(formsPlot1);
-        Margin = new Padding(4, 3, 4, 3);
-        Name = "Form1";
-        StartPosition = FormStartPosition.CenterScreen;
-        Text = "ScottPlot 5 - Windows Forms Sandbox";
+        Name = "MultiplotCollapsed";
+        Text = "MultiplotCollapsed";
         ResumeLayout(false);
     }