+++ src/ScottPlot5/ScottPlot5/Interactivity/MouseAxisManipulation.cs
@@ -44,11 +44,10 @@ public static class MouseAxisManipulation
         }
     }
 
-    public static void DragPan(Plot plot, Pixel mouseDown, Pixel mouseNow)
+    public static void DragPan(Plot plot, Pixel mouseDown, Pixel mouseNow, bool changeOpposingAxesTogether)
     {
         mouseDown = mouseDown.Divide(plot.ScaleFactorF);
         mouseNow = mouseNow.Divide(plot.ScaleFactorF);
-        IPlotControl control = plot.PlotControl ?? throw new NullReferenceException();
 
         float pixelDeltaX = -(mouseNow.X - mouseDown.X);
         float pixelDeltaY = mouseNow.Y - mouseDown.Y;
@@ -62,11 +61,11 @@ public static class MouseAxisManipulation
 
         if (axisUnderMouse is not null)
         {
-            if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
+            if (changeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
             {
                 plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.PanMouse(scaledDeltaX, dataRect.Width));
             }
-            else if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsVertical())
+            else if (changeOpposingAxesTogether && axisUnderMouse.IsVertical())
             {
                 plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.PanMouse(scaledDeltaY, dataRect.Height));
             }
@@ -85,11 +84,10 @@ public static class MouseAxisManipulation
         }
     }
 
-    public static void DragZoom(Plot plot, Pixel mouseDown, Pixel mouseNow)
+    public static void DragZoom(Plot plot, Pixel mouseDown, Pixel mouseNow, bool changeOpposingAxesTogether)
     {
         mouseDown = mouseDown.Divide(plot.ScaleFactorF);
         mouseNow = mouseNow.Divide(plot.ScaleFactorF);
-        IPlotControl control = plot.PlotControl ?? throw new NullReferenceException();
 
         float pixelDeltaX = mouseNow.X - mouseDown.X;
         float pixelDeltaY = -(mouseNow.Y - mouseDown.Y);
@@ -100,11 +98,11 @@ public static class MouseAxisManipulation
 
         if (axisUnderMouse is not null)
         {
-            if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
+            if (changeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
             {
                 plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.ZoomMouseDelta(pixelDeltaX, lastRenderDataRect.Width));
             }
-            else if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsVertical())
+            else if (changeOpposingAxesTogether && axisUnderMouse.IsVertical())
             {
                 plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.ZoomMouseDelta(pixelDeltaY, lastRenderDataRect.Height));
             }