+++ src/ScottPlot5/ScottPlot5/Interactivity/UserActionResponses/MouseDragZoom.cs
@@ -41,19 +41,29 @@ public class MouseDragZoom(MouseButton button) : IUserActionResponse
     /// </summary>
     public double SensitivityY { get; set; } = 1.0;
 
-    public void ResetState(Plot plot)
+    /// <summary>
+    /// If enabled, mouse interactions over a single axis will be applied to all axes with the same orientation.
+    /// </summary>
+    public bool ChangeOpposingAxesTogether { get; set; } = false;
+
+    public void ResetState(IPlotControl plotControl)
     {
         RememberedLimits = null;
         MouseDownPixel = Pixel.NaN;
     }
 
-    public ResponseInfo Execute(Plot plot, IUserAction userInput, KeyboardState keys)
+    public ResponseInfo Execute(IPlotControl plotControl, IUserAction userInput, KeyboardState keys)
     {
         if (userInput is IMouseButtonAction mouseDownAction && mouseDownAction.Button == MouseButton && mouseDownAction.IsPressed)
         {
             MouseDownPixel = mouseDownAction.Pixel;
-            RememberedLimits = new(plot);
-            return new ResponseInfo() { IsPrimary = false };
+            Plot? plot = plotControl.GetPlotAtPixel(mouseDownAction.Pixel);
+
+            if (plot is not null)
+            {
+                RememberedLimits = new(plot);
+                return new ResponseInfo() { IsPrimary = false };
+            }
         }
 
         if (MouseDownPixel == Pixel.NaN)
@@ -61,17 +71,23 @@ public class MouseDragZoom(MouseButton button) : IUserActionResponse
 
         if (userInput is IMouseButtonAction mouseUpAction && mouseUpAction.Button == MouseButton && !mouseUpAction.IsPressed)
         {
-            RememberedLimits?.Recall();
-            ApplyToPlot(plot, MouseDownPixel, mouseUpAction.Pixel, keys);
-            MouseDownPixel = Pixel.NaN;
-            return ResponseInfo.Refresh;
+            if (RememberedLimits is not null)
+            {
+                RememberedLimits.Recall();
+                ApplyToPlot(RememberedLimits.Plot, MouseDownPixel, mouseUpAction.Pixel, keys);
+                MouseDownPixel = Pixel.NaN;
+                return ResponseInfo.Refresh;
+            }
         }
 
         if (userInput is IMouseAction mouseMoveAction)
         {
-            RememberedLimits?.Recall();
-            ApplyToPlot(plot, MouseDownPixel, mouseMoveAction.Pixel, keys);
-            return new ResponseInfo() { RefreshNeeded = true, IsPrimary = true };
+            if (RememberedLimits is not null)
+            {
+                RememberedLimits.Recall();
+                ApplyToPlot(RememberedLimits.Plot, MouseDownPixel, mouseMoveAction.Pixel, keys);
+                return new ResponseInfo() { RefreshNeeded = true, IsPrimary = true };
+            }
         }
 
         return new ResponseInfo() { IsPrimary = true };
@@ -107,6 +123,6 @@ public class MouseDragZoom(MouseButton button) : IUserActionResponse
 
         px2 = new(x2, y2);
 
-        MouseAxisManipulation.DragZoom(plot, px1, px2);
+        MouseAxisManipulation.DragZoom(plot, px1, px2, ChangeOpposingAxesTogether);
     }
 }