+++ src/ScottPlot5/ScottPlot5/Interactivity/UserInputProcessor.cs
@@ -1,9 +1,4 @@
-﻿using ScottPlot.Interactivity.UserActionResponses;
-using System.Reflection.Emit;
-
-namespace ScottPlot.Interactivity;
-
-#pragma warning disable CS0618 // disable obsolete Interaction warning
+﻿namespace ScottPlot.Interactivity;
 
 /// <summary>
 /// This class collects user inputs and performs responses to manipulate a Plot.
@@ -20,7 +15,7 @@ public class UserInputProcessor
     /// <summary>
     /// Tracks which keys are currently pressed
     /// </summary>
-    public readonly KeyboardState KeyState;
+    public readonly KeyboardState KeyState = new();
 
     /// <summary>
     /// Controls whether new events are processed.
@@ -31,11 +26,7 @@ public class UserInputProcessor
         get => _IsEnabled; set
         {
             _IsEnabled = value;
-            if (value)
-            {
-                PlotControl?.Interaction.Disable();
-            }
-            else
+            if (!value)
             {
                 ProcessLostFocus();
             }
@@ -64,7 +55,6 @@ public class UserInputProcessor
     public UserInputProcessor(IPlotControl plotControl)
     {
         PlotControl = plotControl;
-        KeyState = new();
         Reset();
         IsEnabled = true;
     }
@@ -198,7 +188,7 @@ public class UserInputProcessor
     {
         foreach (var response in plotControl.UserInputProcessor.UserActionResponses)
         {
-            response.ResetState(plotControl.Plot);
+            response.ResetState(plotControl);
         }
 
         plotControl.UserInputProcessor.KeyState.Reset();
@@ -236,7 +226,7 @@ public class UserInputProcessor
                     continue;
                 }
 
-                ResponseInfo info = response.Execute(PlotControl.Plot, userAction, KeyState);
+                ResponseInfo info = response.Execute(PlotControl, userAction, KeyState);
                 if (info.RefreshNeeded)
                     refreshNeeded = true;